Split apart models module and update docs
Also starts adding some doctests in to the documentation
