+++ docs/api_reference/broker/models/accounts.rst
@@ -1,0 +1,19 @@
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+Models For Accounts
+===================
+
+This page is for models relating to Accounts.
+
+This has been separated out into a different module internally (and here in the docs) for ease of development
+but is still accessible from ``alpaca.broker.models`` directly.
+
+For example:
+
+>>> from alpaca.broker.models import Account
+
+
+.. automodule:: alpaca.broker.models.accounts
+  :members:
