+++ docs/api_reference/broker/models/cip.rst
@@ -1,0 +1,19 @@
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+Models Relating to CIP Info
+===========================
+
+This page is for models that are focused on CIP info for Accounts.
+
+This has been separated out into a different module internally (and here in the docs) for ease of development
+but is still accessible from ``alpaca.broker.models`` directly.
+
+For example:
+
+>>> from alpaca.broker.models import CIPInfo
+
+
+.. automodule:: alpaca.broker.models.cip
+  :members:
