+++ docs/api_reference/broker/models/requests.rst
@@ -1,0 +1,19 @@
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+Models Representing API Requests
+================================
+
+This page is for the models representing the information that various API requests require/can use.
+
+This has been separated out into a different module internally (and here in the docs) for ease of development
+but is still accessible from ``alpaca.broker.models`` directly.
+
+For example:
+
+>>> from alpaca.broker.models import ListAccountsRequest
+
+
+.. automodule:: alpaca.broker.models.requests
+  :members:
