+++ docs/broker.rst
@@ -5,23 +5,4 @@
 Broker API
 ==========
 
-Broker Models
--------------
-
-.. automodule:: alpaca.broker.models
-   :members:
-
-
-Broker Enums
-------------
-
-.. automodule:: alpaca.broker.enums
-   :members:
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
+TODO: Give overview of broker modules
